ox-html.el: Fix document structure for org-info.js compatibility.
* ox-html.el (org-html-inner-template): Add the document title
here, within the "content" class, as the org-info.js needs it.
(org-html-template): Don't include the document's title here.
(org-html-format-inlinetask-function): Remove wrong example.

Although counter-intuitive and certainly weird, the
<h1 class="title"> string needs to be within the "content"
div, not before it.  We would need the developer of org-info.js
to update org-info.js to put the title outside of "content".