Columnview: Fix the case of #+ content present
* lisp/org-colview.el (org-dblock-write:columnview): Change the
capture of pos to after inserting the original content

The problem is with a block that has content preceding the table.
Upon recreating the content, the `pos' gets set to the beginning of
this content instead of the table.  Later calls to
`org-table-recalculate' or `org-table-align' will fail because the
point is not at a table.

This patch moves the capture of `pos' to right before the insertion of
the table.

TINYCHANGE
Patch by Benjamin Beckwith