Allow the user to control the annotation buffer
* contrib/lisp/org-annotate-file.el (org-annotate-file,
org-annotate-file-show-section): Return the annotation buffer from
`org-annotate-file-show-section` so that the user can decide how to
display it.

When adding an annotation using org-file-annotate, the buffer would
always be opened in the current window.  By not making the buffer
current and instead returning it from `org-annotate-file-show-section`
the user can control how to display it.

For instance, it can be opened in a popup window using popwin:

    (popwin:display-buffer-1 (org-annotate-file-show-section storage-file)
                             :default-config-keywords '(:position :bottom))

The function `org-annotate-file` will still work as before and open
the buffer in the current window.