contrib/lisp/org-passwords.el: New file
* contrib/lisp/org-passwords.el: New file.

The file gives a mode for consulting a passwords database with entries
in org format.  The database opens in Read-Only mode and a timer is
set to close the database.  The file provides functions for making
usernames and passwords available to the facility for pasting text of
the window system (clipboard on X and MS-Windows, pasteboard on
Nextstep/Mac OS, etc.) without inserting them in the kill-ring.  It
also provides functions for generating passwords as a random string of
characters or as a number of random words from a dictionary.